move extractors/smda/__init__.py logic to extractor.py
